Migrations: convert the values to be sent Remove unnecessary console
